+++ lib/alf/errors.rb
@@ -1,5 +1,5 @@
 module Alf
   class Error < StandardError; end
-  class NoSuchDatasetError < Error; end
+  class NoSuchRelvarError < Error; end
   class CoercionError < Error; end
 end