+++ spec/unit/alf-core/relvar/base/test_value.rb
@@ -3,14 +3,15 @@ module Alf
   class Relvar
     describe Base, 'value' do
 
-      let(:database){ examples_database       }
-      let(:name)    { :suppliers              }
-      let(:relvar)  { Base.new(database,name) }
+      let(:db)      { examples_database           }
+      let(:name)    { :suppliers                  }
+      let(:builder) { lambda{|ctx| [{:id => 1}] } }
+      let(:relvar)  { Base.new(db,name,&builder)  }
 
       subject{ relvar.value }
 
-      it 'returns a Relation value' do
-        subject.should be_a(Relation)
+      it 'is obtained through the bulder' do
+        subject.should eq(Relation(:id => 1))
       end
 
     end