+++ CHANGELOG.md
@@ -53,7 +53,7 @@
 
 * The Aggregator class, Summarization type and Summarize operator have been made thread-safe through #happens that now takes a TupleScope instead of a tuple.
 
-* Sequel::Adapter#dataset correctly raises a NoSuchDatasetError if no table can be found.
+* Sequel::Adapter#relvar correctly raises a NoSuchRelvarError if no table can be found.
 
 ## Broken stuff
 
@@ -62,6 +62,8 @@
       Alf.lipsy(db_params).query{ ... } -> Alf.connect(db_params).query{ ... }
 
 * The Environment concept as been removed and split as different abstractions, namely Adapter, Database and Connection. That also means that `environment` readers and writers here have been replaced according to cases. Also, the --env option has been renamed to --db in the shell command. This is a major incompatible change of Alf internals that might break existing code that extends Alf::Environment, Alf::Reader or any subclass.
+
+* Adapter#dataset has been replaced to Adapter#relvar and now serves relation variables instead of pure iterators.
 
 * You now have to explicitely use 'alf show --text > ...' or 'alf --text ... > ' if you don't want ruby hashes to be outputted to output files. This is a consequence of tty detection that ensures a better shell experience.